requirements.txt
absl-py==0.11.0
altgraph==0.17
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pefile==2019.4.18
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
